Updating Singleton Instance, as well as CI (#31)
* update spec for interop

* Updated Singleton instance implementation.

The correct call of the singleton instance implemented.
All subsequent singleton classes that are implemented hence forth are implemented as:

```python

class MySingletonClass(Singleton):
	def __init__():
		pass
```

* Singleton update for concurrency

Singleton decorator has been updated, and tested.
Use of this singleton class decorator can be done as follows.

```python
import singly

@Singleton
class TestClass():
	def __init__(self):
		#something
		pass

```

* Updating Docmentation

This updates the documentation for the singleton implementation
It also demonstrates the implemnetation.

* Update README.md

* Update testing scripts

Testing scripts required for integration.

* Added testing scripts along with integration

Integration is through the travis.yml script that is parsed. Testing is for the Singleton

* Changed name of singly

Changed the name of singly to singleton. This enhances the development, and usage for developers

* Updated documentation for integration

* Updated package specfiics

* Updated Singleton class naming conventions with CI

Singleton class name change for decorator,
Continuous Integration support for future builds

* Updated travis file for reqs